changed totalPages value to double

+++ src/main/java/com/example/demo/repository/ClientInfoRepository.java
@@ -21,5 +21,5 @@ public interface ClientInfoRepository {
 	public void insertClientInfo(@BindBean ClientInfo clientInfo);
 	
 	@SqlQuery("SELECT COUNT(*) FROM ClientInfo")
-	public int getTotalPages();
+	public double getTotalPages();
 }
