+++ src/main/java/com/example/demo/services/ClientInfoService.java
@@ -7,5 +7,5 @@ import com.example.demo.model.ClientInfo;
 public interface ClientInfoService {
 	public List<ClientInfo> getClientInfo(int id);
 	public void insertClientInfo(ClientInfo clientInfo);
-	public int getTotalPages();
+	public double getTotalPages();
 }
